e_menu_backend/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-extensions==3.1.0
django-filter==2.4.0
django-heroku==0.3.1
django-hosts==4.0
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
gunicorn==20.0.4
install==1.3.4
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
